Refactored screen savers as creators
Necessary to emphasize that these objects create and/or save images as attachments for screen records in the database. This also helps align these objects more closely with `Terminus::Aspects::Screens::Creator` as it's the primary object that delegates to these creators.

Issue: 194
Milestone: patch
