+++ app/aspects/screens/creators/encoded.rb
@@ -1,19 +1,19 @@
 # frozen_string_literal: true
 
+require "base64"
 require "dry/monads"
-require "mini_magick"
 
 module Terminus
   module Aspects
     module Screens
-      module Savers
-        # Saves URI as a processed image.
-        class Unprocessed
-          include Deps["aspects.screens.converter", repository: "repositories.screen"]
+      module Creators
+        # Creates screen record with image attachment from decoded image data.
+        class Encoded
           include Dry::Monads[:result]
+          include Deps["aspects.screens.converter", repository: "repositories.screen"]
 
-          def initialize(client: MiniMagick::Image, struct: Terminus::Structs::Screen.new, **)
-            @client = client
+          def initialize(decoder: Base64, struct: Terminus::Structs::Screen.new, **)
+            @decoder = decoder
             @struct = struct
             super(**)
           end
@@ -22,15 +22,14 @@ module Terminus
 
           private
 
-          attr_reader :client, :struct
+          attr_reader :decoder, :struct
 
           def process payload, directory
             input_path = Pathname(directory).join "input.png"
 
-            client.open(payload.content)
-                  .write(input_path)
-                  .then { convert payload.model, input_path, directory.join(payload.filename) }
-                  .bind { |path| save payload, path }
+            input_path.binwrite(decoder.strict_decode64(payload.content))
+                      .then { convert payload.model, input_path, directory.join(payload.filename) }
+                      .bind { |path| save payload, path }
           end
 
           def convert model, input_path, output_path
