+++ app/aspects/screens/creators/payload.rb
@@ -4,8 +4,8 @@
 module Terminus
   module Aspects
     module Screens
-      module Savers
-        # Defines payload for use in saving screens.
+      module Creators
+        # Defines payload for use in creating screens.
         Payload = Data.define :model, :name, :label, :content do
           def attributes = {model_id: model.id, name:, label:}
