+++ spec/app/aspects/screens/creators/html_spec.rb
@@ -2,20 +2,32 @@
 
 require "hanami_helper"
 
-RSpec.describe Terminus::Aspects::Screens::Savers::Encoded, :db do
+RSpec.describe Terminus::Aspects::Screens::Creators::HTML, :db do
   subject(:saver) { described_class.new }
 
   describe "#call" do
-    let :payload do
-      Terminus::Aspects::Screens::Savers::Payload[
-        model:,
-        name: "test",
-        label: "Test",
-        content: Base64.strict_encode64(SPEC_ROOT.join("support/fixtures/test.png").read)
-      ]
+    let(:model) { Factory[:model] }
+
+    let :content do
+      <<~CONTENT
+        <html>
+          <head>
+            <style>
+              color: black;
+              background-color: black;
+            </style>
+          </head>
+
+          <body>
+            <h1>Test</h1>
+          </body>
+        </html>
+      CONTENT
     end
 
-    let(:model) { Factory[:model] }
+    let :payload do
+      Terminus::Aspects::Screens::Creators::Payload[model:, name: "test", label: "Test", content:]
+    end
 
     it "answers screen" do
       result = saver.call payload
@@ -39,6 +51,12 @@ RSpec.describe Terminus::Aspects::Screens::Savers::Encoded, :db do
     it "answers failure with database error" do
       result = saver.call payload.with(name: nil)
       expect(result.failure).to match(/null value/)
+    end
+  end
+
+  describe "#inspect" do
+    it "only displays the sanitizer class" do
+      expect(saver.inspect).to include("@sanitizer=Terminus::Sanitizer")
     end
   end
 end
